requirements.txt
torch>=1.7.0
torchtext==0.8.0
spacy>=3.0.0
numpy>=1.19.0
tqdm>=4.48.0
pyyaml>=5.4.0
matplotlib>=3.3.0
tensorboard>=2.4.0 